+++ docs/ChangeLog.md
@@ -2,6 +2,8 @@
 - Removed accidental 'or short phrases' from AI term translation
 - Fixed Audio icon not showing in books list
 - Fixed weekly stats to show only 7 days instead of 8
+- Fixed RTL formating
+- Fixed more audio player issues
 
 # New Feature
 -